chore: move aws logic
