+++ packages/domains-api/src/route53/create_or_update_hosted_zone.ts
@@ -1,4 +1,4 @@
-import logger from '../logger';
+import logger from '../config/logger';
 import { HostedZone, DelegationSet } from './types';
 
 type Args = {
