+++ packages/domains-api/src/route53/delete_hosted_zone.ts
@@ -1,4 +1,4 @@
-import logger from '../logger';
+import logger from '../config/logger';
 
 export default (route53: any) => async ({ hostedZoneId }: any): Promise<void> => {
   const result = await route53.deleteHostedZone({ Id: hostedZoneId });
